requirements.txt
arcade
shapely
pyyaml 
numpy
matplotlib 
zmq
icecream
requests
